вьюмодели вместо дикшинари

+++ ViperCoreDataDZ/ViperCoreDataDZ/UserListModule/InteractorUser.swift
@@ -9,27 +9,30 @@
 import UIKit
 import CoreData
 class InteractorUser: NSObject, InteractorUserListProtocol {
-	
+	var users = [VDUserSpecial]()
 	func updateDataBase() {
-    	VDDataManager.sharedManager.updateUserBD()
-    	VDDataManager.sharedManager.updateCourseBD()
+		VDDataManager.sharedManager.updateUserBD()
+		VDDataManager.sharedManager.updateCourseBD()
 	}
-	func createDict() -> [String: [Special]]{
-    	return ["Users":VDUserSpecial.users]
+	func returnData() -> [VDUserSpecial]{
+		users = VDUserSpecial.users
+		users.sortingBy(parameters: ["firstName", "lastName"])
+		return users
 	}
-	func getData(updateCell:(_ data:[String:[Special]])->()) {
-    	
-    	updateCell(createDict())
+
+	func getData(getUsers:([VDUserSpecial])->()) {
+		getUsers(returnData())
 	}
-	func addEmptyUser() -> NSManagedObjectID {
-    	return VDDataManager.sharedManager.addEmptyUser()
+	func addEmptyUser() -> VDUserSpecial {
+		return VDDataManager.sharedManager.addEmptyUser()
 	}
 	func getUserByID( id: NSManagedObjectID) -> VDUserSpecial{
-    	return VDUserSpecial.users[VDUserSpecial.getUserIndexByID(id: id)!]
+		return VDUserSpecial.users[VDUserSpecial.getUserIndexByID(id: id)!]
 	}
-	func deleteObjectWithIDFromDB(id: NSManagedObjectID) {
-    	VDDataManager.sharedManager.deleteByID(id: id)
+	func deleteObjectFromDB(object: VDUserSpecial) {
+		guard let id = object.ID else {
+			return
+		}
+		VDDataManager.sharedManager.deleteByID(id: id)
 	}
-	
-	
 }
